requirements.txt
streamlit==1.31.0
boto3==1.34.34
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.8
botocore==1.34.34
google-generativeai>=0.8.0